Text.Pretty.Simple.Internal.Color

Description

Documentation

data ColorOptions

These options are for colorizing the output of functions like pPrint.

If you don't want to use a color for one of the options, use colorNull.

Constructors

ColorOptions

Fields

colorQuote :: Style

Color to use for quote characters (") around strings.

colorString :: Style

Color to use for strings.

colorError :: Style

Color for errors, e.g. unmatched brackets.

colorNum :: Style

Color to use for numbers.

colorRainbowParens :: [Style]

A list of colors to use for rainbow parenthesis output. Use '[]' if you don't want rainbow parenthesis. Use just a single item if you want all the rainbow parenthesis to be colored the same.

defaultColorOptionsDarkBg :: ColorOptions

Default color options for use on a dark background.

defaultColorOptionsLightBg :: ColorOptions

Default color options for use on a light background.

colorNull :: Style

No styling.

data Style

Ways to style terminal output.

Constructors

Style

Fields

styleColor :: Maybe (Color, Intensity)

styleBold :: Bool

styleItalic :: Bool

styleUnderlined :: Bool

color :: Intensity -> Color -> Style

colorBold :: Intensity -> Color -> Style

convertStyle :: Style -> AnsiStyle